Started serial version of vdj
Mainly, everything is reimplemented so that it does not have
to load all the data into memory, but iterates and operates
over the VDJXML files.

The focus will also be to write a set of scripts like BLAST,
exonerate, or whatever else that operates on big files and
spits out other big files, rather than working in memory.

__init__parallel.py is the old version of __init__.py for use
as a reference temporarily before it gets totally eliminated.